Moved component css to styles folder
