+++ pages/signup.tsx
@@ -1,5 +1,5 @@
 import { NextPage } from "next";
-import Signin from "../components/singin/signin";
+import Signin from "../components/signin/signin";
 
 const SignupPage: NextPage = () => {
     return(
